refactor: conversation component and remove unused code
