+++ frontend/src/views/patient/components/Skeleton.jsx
@@ -1,9 +1,11 @@
 const Skeleton = () => {
   return (
     <div className="flex flex-col gap-4 w-full">
-      <div className="skeleton h-4 w-full"></div>
-      <div className="skeleton h-4 w-full"></div>
-      <div className="skeleton h-4 w-1/2"></div>
+      <div className="skeleton h-3 w-full"></div>
+      <div className="skeleton h-3 w-full"></div>
+      <div className="skeleton h-3 w-full"></div>
+      <div className="skeleton h-3 w-full"></div>
+      <div className="skeleton h-3 w-1/2"></div>
     </div>
   );
 };
